+++ client/i18n.js
@@ -1,6 +1,6 @@
 import i18next from 'i18next'
 import { initReactI18next } from 'react-i18next'
-import * as locales from "../locales/index.ts"
+import * as locales from "../localization/locales/index.ts"
 
 i18next
     .use(initReactI18next)
