--- locales/en.json
@@ -1,34 +1,0 @@
-{
-    "open-explore-tables": "Open & explore tables",
-    "detect-errors-generate-report": "Detect errors and generate a report",
-    "save-download-work": "Save & download your work",
-    "welcome-to-ODE": "Welcome to the Open Data Editor!",
-    "welcomebanner-description": "The ODE helps data practitioners with no coding skills to explore tabular data and detect errors in an easier way. Advanced users can also edit metadata and publish their work.<br /><br /> The OKFN team aims to add other data formats in the future.",
-    "link-check-blog": "Check our blog for updates",
-    "get-started": "Get Started",
-    "welcomebanner-dont-show-again": "Don't show this screen on the next launch",
-    "powered-by-frictionless": "This project is open source and powered by",
-    "how-ODE-handles-your-files": "How the ODE handles your imported files",
-    "openlocationbanner-description" : "The tool makes a copy of your selections into the ODE folder on your Computer. Any changes will be made to these copies, while the original data remains unchanged.",
-    "openlocation-dont-show-again": "Don't show next time",
-    "okay": "Okay",
-    "next": "Next",
-    "back": "Back",
-    "alt-open-file-location": "Open File Location Slide",
-    "open-file-location": "Open File Location",
-    "upload-your-data": "Upload your data",
-    "create-an-issue": "create an issue",
-    "sharing-contents-if-possible": "sharing the file contents <small>(if possible)</small>",
-    "ODE-supports-CSV-Excel-files" : "The ODE supports Excel & csv files",
-    "links-online-tables" : "You can also add links to online tables",
-    "create-folder": "Create folder",
-    "user-guide": "User guide",
-    "report-an-issue": "Report an issue",
-    "from-your-computer": "From your computer",
-    "add-external-data": "Add external data",
-    "add-folders": "Add one or more folders",
-    "add-xsl-or-csv-files": "Add one or more Excel or csv files",
-    "link-external-table": "Link to the external table:",
-    "add": "Add",
-    "enter-or-paste-url": "Enter or paste URL"
-}
