+++ localization/locales/en.json
@@ -1,0 +1,96 @@
+{
+    "open-explore-tables": "Open & explore tables",
+    "detect-errors-generate-report": "Detect errors and generate a report",
+    "save-download-work": "Save & download your work",
+    "welcome-to-ODE": "Welcome to the Open Data Editor!",
+    "welcomebanner-description": "The ODE helps data practitioners with no coding skills to explore tabular data and detect errors in an easier way. Advanced users can also edit metadata and publish their work.<br /><br /> The OKFN team aims to add other data formats in the future.",
+    "link-check-blog": "Check our blog for updates",
+    "get-started": "Get started",
+    "welcomebanner-dont-show-again": "Don't show this screen on the next launch",
+    "powered-by-frictionless": "This project is open source and powered by",
+    "how-ODE-handles-your-files": "How the ODE handles your imported files",
+    "openlocationbanner-description" : "The tool makes a copy of your selections into the ODE folder on your Computer. Any changes will be made to these copies, while the original data remains unchanged.",
+    "openlocation-dont-show-again": "Don't show next time",
+    "okay": "Okay",
+    "ok": "Ok",
+    "next": "Next",
+    "back": "Back",
+    "alt-open-file-location": "Open File Location Slide",
+    "open-file-location": "Open File Location",
+    "upload-your-data": "Upload your data",
+    "create-an-issue": "create an issue",
+    "sharing-contents-if-possible": "sharing the file contents <small>(if possible)</small>",
+    "ODE-supports-CSV-Excel-files" : "The ODE supports Excel & csv files",
+    "links-online-tables" : "You can also add links to online tables",
+    "create-folder": "Create folder",
+    "create-new-folder": "Create new folder",
+    "Create": "Create",
+    "user-guide": "User guide",
+    "report-an-issue": "Report an issue",
+    "from-your-computer": "From your computer",
+    "add-external-data": "Add external data",
+    "add-folders": "Add one or more folders",
+    "add-xsl-or-csv-files": "Add one or more Excel or csv files",
+    "link-external-table": "Link to the external table:",
+    "add": "Add",
+    "enter-or-paste-url": "Enter or paste URL",
+    "error-url-blank": "The URL is blank",
+    "error-url-not-valid": "The URL is not valid",
+    "error-google-sheets-address-invalid": "The Google Sheets URL is not valid or the table is not publically available",
+    "error-url-not-table": "The URL is not associated with a table",
+    "error-file-size-exceeds-10mb": "The total size of the files exceeds 10MB. This operation might take some time...",
+    "checking-errors": "Checking errors",
+    "rename": "Rename",
+    "file": "file",
+    "folder": "folder",
+    "delete": "Delete",
+    "open-fileorfolder-location": "Open {{fileOrFolder}} Location",
+    "context-menu-delete-description":"Only removes this element from the ODE folder",
+    "context-menu-openlocation-description": "The ODE folder where this {{fileOrFolder}} exists",
+    "metadata": "Metadata",
+    "errors-report": "Errors report",
+    "source": "Source",
+    "chat-AI": "Chat AI",
+    "undo": "Undo",
+    "redo": "Redo",
+    "new-file": "New file",
+    "new-folder": "New folder",
+    "external-data": "External data",
+    "publish": "Publish",
+    "edit": "Edit",
+    "view": "View",
+    "toggle-metadata": "Toggle Metadata Panel",
+    "toggle-errors": "Toggle Errors Panel",
+    "toggle-source": "Toggle Source Panel",
+    "help": "Help",
+    "ODE-user-guide": "ODE User guide",
+    "name-new-folder": "Name of the new folder",
+    "save": "Save",
+    "name-new-filefolder": "Name of new {{ fileOrFolder }}",
+    "rename-filefolder": "Rename {{ fileOrFolder }}",
+    "no": "No",
+    "delete-fileFolder": "Delete {{ fileOrFolder }}",
+    "are-you-sure-delete-elements": "Are you sure you want to delete these elements?",
+    "are-you-sure-delete-filefolder": "Are you sure you want to delete this {{ fileOrFolder }}?",
+    "cancel": "Cancel",
+    "dialog": "Dialog",
+    "loading": "Loading",
+    "confirm": "Confirm",
+    "close": "close",
+    "save-changes": "Save changes",
+    "unsaved-changes": "Unsaved changes",
+    "discard": "Discard",
+    "unsaved-changes-dialog-description": "There are unsaved changes. Please, click save or cancel.",
+    "publish-dataset": "Publish dataset",
+    "dataset": "Dataset",
+    "base-url": "Base Url",
+    "allow-update": "Allow Update",
+    "api-key": "API Key",
+    "user": "User",
+    "repo": "Repo",
+    "email": "Email",
+    "description": "Description",
+    "author": "Author",
+    "failed-open-project": "<strong>Failed to open the project</strong>. Please",
+    "preview-not-available": "Preview is not available for this file format"
+}
